Use 32 bits insted of 8 for warning status.
Warning status can be in fact represented with only 2 bits, but
since it's engaged in arithmetic operations it gets promoted and
special precautions are needed. Much easier to ure 32 bits:
warnings are nowhere near performance/memory bottleneck.

+++ re2c/src/conf/warn.cc
@@ -9,9 +9,9 @@ namespace re2c {
 
 Warn warn;
 
-const uint8_t Warn::SILENT  = 0;
-const uint8_t Warn::WARNING = 1u << 0;
-const uint8_t Warn::ERROR   = 1u << 1;
+const uint32_t Warn::SILENT  = 0;
+const uint32_t Warn::WARNING = 1u << 0;
+const uint32_t Warn::ERROR   = 1u << 1;
 
 const char * Warn::names [TYPES] =
 {
